Allow Dateish *day* parameter to be a Callable / Whatever (#4808)
When creating a Date or DateTime object, you sometimes want to
specify the day from the end of the month in the given year.
You can do that now by for x, similar to end-of-array indexing:

    my $date = Date.today;
    say $date.clone(day => $date.days-in-month - x);

This is nice, but it means needing to create a Date/DateTime object
before getting your intended Date/DateTime object.

With this commit, you can also specify a Callable for the day parameter:

    say Date.new(2022, 3, *);  # 2022-03-31

where the * represents the number of days in the month in that year.  Or:

    say Date.new(2022, 3, *-2);  # 2022-03-29

for two days before the last day of a month.