Fix resources of the parent repo not found when using Staging
Resources are loaded lazily, so we pick up changes in the repository's location.
Indeed, this is exactly what happens when a dist gets installed via the Staging
repsitory: the files will end up in a different location than they were
installed to. The lookup uses the name of the repository, the dist id and the
resource's name. Since the Staging repo shadows its parent, it will end up
handling all requests, even for dists that are not installed into Staging, but
are already in its parent.

The source-file method already took care of this by delegating failed lookups to
the parent. We need to do the same for resources.

Fixes GH #4819