CURI: pre-compile longest names first
The idea being that the longest names have fewer dependenencies
(in general).  Also make sure that we only sort the entries once.

It would probably be better if we could specify an order in which
modules should be pre-compiled.  But alas, the current META6.json
format is a hash, so without order.  And we need some sort of order
to get reproducible builds.