Move CUR::Staging into core
Yes, it is only needed for very specific applications.  But it is
a very small module, and it simplifies the various install scripts
significantly (most notable "install-core-dist.raku").  And it won't
break code that actually do a 'use CompUnit::Repository::Staging".

And it reduces the core setting compilation with about a second.
Which, on a normal core development day, I'm doing a *lot*.