Only assign to caller's $/ during a regex subsititution if it's writable
https://github.com/rakudo/rakudo/pull/4807 removed a `try` before the
assignment that was added in 14c1a6f58a73ed0ab92f3fb379e332f682afa916
because I couldn't find an example that needed it. However, that broke
YAMLish (it was doing a subsititution in an actions method). However, a
`try` is overkill here, and we can just make the cheaper check for
writability.