Re-imagine CompUnit::Repository::Spec.from-string
This was using regular expressions and meta ops, for code that is
really pretty simple.  Since this is being called multiple times
for *every* pre-compilation, it seems to make sense to streamline
this code using nqp::ops and without regexes and meta-ops.

Underlying reason was that another upcoming fix broke because
meta-ops were being used *before* the Rakudo::Metaops mapper had
been initialized.