Add :precompile = True flag to CURI.install (#4826)
The idea behind this is to allow modules to just be installed and
have the runtime decide when it should (re-)precompile any compilation
units that were not precompiled yet.  This would e.g. allow the
DateTime::TimeZone distribution to be installed within a second,
rather than on 4+ minutes.  And only precompile those timezones that
are actually used at runtime.

Also add the --/precompile flag to install-dist.raku