Re-imagine CompUnit::Repository::Spec.from-string (#4837)
This was using regular expressions and meta ops, for code that is
really pretty simple (apart from the option parsing).  Since this
is being called multiple times for *every* pre-compilation, it seems
to make sense to streamline this code and with fewer regexes and
no meta-ops.

Underlying reason was that another upcoming fix broke because
meta-ops were being used *before* the Rakudo::Metaops mapper had
been initialized.

Also be more allowing for values.