Attach position information to parse errors
So far, the RakuAST-based compiler hasn't been reporting the location of
syntax errors. The way this is factored in the current compiler mixes
the concerns of assembling the exception object with extracting various
bits and pieces that it's the parser's job to know about. In RakuAST we
need something a bit cleaner, and less coupled to any particular parser.
Thus, keep `build-exception` on the resolver to just taking a set of
named arguments, and construct those parser-provided ones in the parser
itself.

Ultimately this won't suffice, because we need to be able to attach
position information for semantic errors (those determined inside of the
AST nodes at CHECK time) too.