Have package body consistently be a block
It's typed that way already in the AST definition, although we don't
yet enforce these type constraints. We could in principle arrange for
something a bit like RakuAST::Blorst (that doesn't apply here since it's
a block or statement list). However, it seems more expedient to define a
unit package (or unit sub main) as sugar for a package or sub with a
block whose content extends to the end of the current file.