Implement =finish support in RakuAST
And use it in the RakuAST-based compiler. This matches the current
implementation's behavior of giving Mu if there's no finish content,
although it'd seem at least a warning could be nice (could be done by
implementing RakuAST::CheckTime in the Finish code).

Didn't constrain the finish content to Mu in the AST node, since it
occurs to me that folks could in principle write modules that add a
custom compiler pass that parses the `=finish` content in some way
(for example `use Finish::JSON` would parse the `=finish` content as
JSON, and `$=finish` would thus give the parsed JSON back).