Make IO::Path.mkdir 2x as fast if dir already exists
If the dir does not exist, it is only one cheap existence check extra,
which dwarves what is needed to actually make a directory.  And since
IO::Path.mkdir returns the invocant, one would be inclined to write
"foo".IO.mkdir.add("bar") to create a file "bar" in a "foo" directory.
This is now significantly faster if the directory already exists.