Streamline CompUnit::DependencySpecification logic
Specifically with regards to the xxx-matcher logic.  Before, one
had to explicitely specify True to indicate no interest in matching
that matcher (which would essentially match against True, and thus
always pass).  And if something *was* specified for "version" or
"api", then at several places a Version object was made to get the
proper matching semantics.  This logic has now moved into the
CompUnit::DependencySpecification logic.

This also caches the stringification of a CompUnit::DependencySpecification
object, as it is relatively expensive.

This also adapts all calls to CompUnit::DependencySpecification in
the core.

Also add some internal documentation as to the functioning of it all.