Streamline CURI.files
Mostly about not creating intermediate variables, and making sure
that we only return meta info for existing files to begin with,
instead of first returning the meta and *then* check if the file
exists.

Also add a BIND-KEY method to the MetaAssigner role so that we can
transparently bind to something that looks like a hash.