Shortcut .map/.grep/.first on Empty
The empty Slip (aka Empty) is returned in many cases.  This commit
adds Slip candidates for .map / .grep / .first that will check whether
the invocant is indeed the special Empty case, and then just return
the appropriate thing (Empty in case of .map / .grep, Nil in case of
.first).  In all other cases, run the normal handling of .map / .grep
/ .first.

Makes .map on the empty Slip about 4x as fast.