Streamline CURI.resource
- use native hash for caching, allowing for ifnull usage
- make sure that caching distribution metas is thread-safe
- remove check for repo versions < 2

Note that this removes the check for repo's of repository level 0
and 1.  Level 0 is until February 2016, and level 1 is until May 2016.
It seems *highly* unlikely that anybody would still be having a repo
around from that era that has not yet been upgraded.

Level 0 can **also** occur on a completely new, empty repository.
But then there aren't any resources to produce anyway, so removing
the check in that case, is not an issue.