Add a script to empty out all precomp in the repo chain
Intended to be used by core developers who research / maintain
the precompilation process in Rakudo.

Please note that removing all precomp files does *not* change
functionality of an installation: it will merely free up a lot of
diskspace of you're a core developer compiling the setting over and
over again.  And it will slow down every first load of a compilation
unit.