Fix trait stumbling over unresolved types in signature
We have to fully resolve a routine's signature before applying traits as the
latter may want to access e.g. parameter types. However, naively doing so
revealed a timing problem. Signatures and parameters tuned their behaviour
according to whether they were attached to a (methodish) routine. This broke
as at the time we're applying traits, the signature is not attached yet.

Follow the "tell, don't ask" principle of object oriented design and instead
of signatures and parameters querying their owner, have the owner tell them
how to behave via the new set-is-on-method and set-default-type methods.