Fix topic var and placeholder signature semantics
For { :$^foo } we would end up generating an optional `$_` parameter
instead of just binding the outer value into the block's `$_`. This was
observable because the `$_` parameter that was mistakenly generated was
optional but code gen'd before the required `$^foo`, which in turn blew
up bytecode compilation. The previous approach failed to work due to
timing: the implicit delcarations are needed earlier. Even if it wasn't
for that timing issue, however, blocks should still get a `$_` bound to
the outer scope's one.