Bump NQP for a change/optimization to how bytecode is written to files
The buffer for each frame is now pre-allocated and each frame is written
to the file directly, instead of building them all up into one large
buffer. Given the speed of modern SSDs, this is the same as or only
slightly slower than doing it all at once and saves on total memory
used.