Streamline ModuleLoader.load_setting
By the time this gets called, the setting name has already been
normalized, so we can throw out the check for 'NULL.c' and normalizing
(again).

Make the lookup for previous setting a local hash, so that we do not
have to do a lookup in a hash, call a method on that, then do a lookup
and *then* do the setting name lookup on that.

Reduce the number of dynamic var lookups.

Simplify RAKUDO_MODULE_DEBUG output.

Correct some comments.