Restore original REPL.repl-eval method
Apparently, some modules in the ecosystem call this method (which I
would consider an implementation detail as it is untested and undocumented)
directly.  This solves some of the error messages as seen in Blin
run #1 for 2022.04, but not all.  Here's hoping another Blin run can
bisect the commit that causes the remaining issues.