Remove dead code, japhb++ for noticing
Originally I planned to report the actual size of the blocks freed
but at least on (my) MacOS the reported block size and the number
of blocks freed, was at least a factor 8 too large (which could
very well be it reporting a block size of 4096, but actually working
with 512 byte blocks).  So I returned to just report the sum of the
file sizes of the files removed.  And forgot to remove the code that
tried to calculate the raw size freed.