Reinstate CATCH versus LEAVE
Alhough the code with LEAVE is much simpler and clearer, it *does*
impose the overhead of having to execute a phaser on the most common
execution path.  And since both a CATCH block or a LEAVE phaser will
inhibit inlining of a block, from a performance point of view a
CATCH is better, as that will only be executed when something actually
goes wrong (which hopefully is not the most common execution path).