Put CATCH blocks at the start of the scope
CATCH blocks affect the **entire** scope they're in.  So they should
be as close to the start of the scope they're in, so that you realize
that they can fire well before you think they'd might fire.

This did not change the Rakudo::Internals::JSON class, as that is
being kept in sync with JSON::Fast.