Make &[max] and &[min] return RHS for ties
&[max] and &[min] both have List associativity, which means that

   $a max $b max $c

should be the same as

   max $a, $b, $c

however, in the case of ties, this was not true: &max would return the
first argument with that value, whereas &[max] would return its LHS.
For example, before this commit:

   max 0, False # OUTPUT: «0»
   0  max False # OUTPUT: «False»

These two lines should return the same value, and &max's behavior
seems to be preferable under logic similar to stability in sorting.

This commit changes &[max] so that it returns its RHS in the case of
ties.  It also makes corresponding changes to &[max] candidates where
ties between different values aren't possible (e.g., if $a and $b are
both native ints) to avoid confusing differences in adjacent lines of
code.