Make handles on attribute of type object more natural
Instead of dying with: "Cannot look up attributes in a Foo type
object".  The error message is confusing, take this example:

    class A { has Int $.foo handles <Str> }; (my A $a).Str

To the casual observer, calling .Str on a type object should just
work: there are no attributes in play.  *IF* we want to have that
behaviour, then we should probably mention the name of the actual
attribute.

But I think it is better to assume a decontainerized uninitialized
attribute in that case to more mimic the actual situation.  That
changes it into a warning in this case:

    Use of uninitialized value of type Int in string context.
    Methods .^name, .raku, .gist, or .say can be used to stringify it to something meaningful.

It will definitely make using "handles" much more user friendly.