Use a specific exception for error and optimize code
Since we know the target and all the attributes we added so far, there
is no need to re-scan its attributes for each applied role. Just keep
track of them in a temporary hash.

In order to know the source of each attribute entries on the hash are
records with two attributes: the attribute object itself, and the role
where it has been picked from.

Eventually, we now report all parties involved: the target role, and
both source roles where conflicting attributes are coming from.