Try to resolve races in CURFS
The primary problem solved is concurrent access to `%!loaded` where
separate locks on read and write in method `need` did not provide
bulletproof protection from races. This commit proposes different
approach where the first request for a `$spec` installs a `Promise` and
all concurrent requests for the same spec are awaiting for it to be kept
before returning the resulting `CompUnit` instance.

To avoid always calling `await` on values of `%!loaded`, kept promises
are placed with actual `CompUnit`s.

Due to changes in logic, different locks are now needed for `%!loaded`
and `%!seen`.

Additionally, the way of initializing lazy attributes has been changed
too. Instead of insecure `//=` or alike, `cas` is used initialize the
attributes. And to avoid calling `cas` on every accessor invocation,
corresponding attribute is first tested for definedness. So, eventual
approach is:

    ⚛$!attribute // cas $!attribute, { $_ // <initializer> }