Reconsider %REQUIRE_SYMBOLS
First, change the name to `%?REQUIRE-SYMBOLS`. This makes sense for
reducing the chances of conflicting with user's code lexical. Also,
since this is a true compile-time symbol it deserves the `?` twigil.

Second, make it a `Stash`. For now this mostly makes sense
ideologically, as it is a symbol container in first place. But also,
since some steps were already made to reduce chances of races in `Stash`
code, this helps in making multi-threaded code more robust.