Simplify setting up standard handles
We don't have any thread-safety issues with setting up the standard
handles, and we always use utf8-encoding for them.  So we can set them
up without thread-safety as they are all part of running the mainline.

This commit:
- adds an implementation-detail method to get the utf8 encoding object
  in an thread-unsafe manner
- caches the encoder / decoder objects, so we don't need to look them
  up again and again for the 3 standard handles
- thus removes the need for 3 Lock.protect cases

This appears to scrape off a few msecs from each rakudo startup