Make some Stash operations safer on concurrency
Stash modifications based on key binding are now lock-protected and
consist of three steps: storage cloning -> binding -> storage
replacement. The last step is supposed to be atomic. But due to problems
with MoarVM serialization it is currently unsafe vanilla
`nqp::bindattr`.

- Added own BIND-KEY method
- `package_at_key` stubs a missing package under lock protection
- `merge-symbols` is safe wrapper for `ModuleLoader.merge_globals`.

With regard to the last item, the current approach is for
`merge_globals` to call `merge-symbols` on any `Stash` target it
encounters. To prevent deep recursion and reclaim some of the lost
performance, `merge-symbols` passes cloned symbol hash back into
`merge_globals`.