Give generics accepts_type typechecking semantics
Set the typecheck mode to 2 for consideration of accepts_type should a
typecheck fail, while ignoring the type cache (no identity is valid; Mu
typechecks without our help).  We'll refer to this by its nqp::const
name for readability's sake.

A generic may appear in a typecheck of a signature or an incomplete
parameterization of a role. A generic should never type_check, but
should always accept another object. This gives generics complete
support for ACCEPTS, which comes with a couple benefits:

- Generic roles in a curry's RTL can be checked now
- Generic parameters used as types in a signature can be checked now