Give Signature.ACCEPTS support for generics
A generic parameter has a type associated with it (`Mu` by default). If
we build a map of inferrable typings based on this, we can rule out
signatures that should *not* pass, but do because of
`Metamodel::GenericHOW.accepts_type` (roles can depend on its truthy
nature). We leave the flunky return typecheck alone for now.