Allow :from<Raku> in identities
This allows identities to be specified with :from<Raku> so that we
can start migrating from the current default of "Perl6".  Note that
this does **not** make "Raku" the default yet, as some modules in
the ecosystem might still check for "Perl6" explicitely.

Creation of CompUnit and CompUnit::DependencySpecification objects
should be next (as well as their creation from World)