Reparameterize to instantiate curried roles + grouped role candidates
Because this should also produce a new curry, but may be capable of
caching its parameters, as is the case with parametric role groups.
This makes generics in doees and parents equivalent to those
parameterized by the user wherever possible, allowing for type_check to
succeed before accepts_type in the case of a curried role group.