Fix segfault due to missing description on CompUnit
If getscdesc is called on a comp unit without description, it puts a NULL into
a string register. This may lead to segfaults or hash key lookup failures.
While this is clearly a bug in the VM, it also shows that setting the
description was missing from the RakuAST implementation.