Remove counterproductive fallback from ModuleLoader
Perl6::ModuleLoader needs to restore the "GLOBAL" HLL symbol after loading a
module (in case it got changed). However, if no GLOBAL had been set, it would
get a VMNull and then fall back to NQPMu. In that case, it would "restore" an
NQPMu that wasn't there before. This is bad because the generated code for
installing a GLOBAL object only does so if getcurhllsym returned a VMNull.
The result is that we never install a symbol in the first place.

This issue was hidden before because the compiler would always bind the hll
symbol and RakuAST doesn't (yet). It would for example still be an issue for
precompiled scripts.