Make Supply use $*AWAITER.await
It was using the sub `await` which is defaulting to 6.c versions for the
core and using blocking awaiter. Switching to `$*AWAITER.await` makes it
possible to use non-blocking ThreadPoolAwaiter when available.

Proves helpful under some circumstances when otherwise `whenever` may
block due to no available threads left in the thread pool. Mind the
exception when used under acquired `Lock` (`nqp::lock`).