Replace Lock with Lock::Soft in CU and Stash code
This fixes problems with race conditions in `require` on multiple
levels: from blocked threads when too many simultaneous requests are
issued to `CompUnit::PrecompilationStore::FileSystem`; to `Stash`
falsely returning `VMNull` or `Any` when requested for a loaded module.

Due to a bug in handling `LEAVE` phaser, when used in a method, values
returned with `return` from within a block passed into the `protect`
method of `Lock::Async` and `Lock::Soft` are getting replaced with
various garbage. This doesn't happen with `Lock` though. This is why
many chunks of this commit are about moving `return`/`fail` from the
protected blocks.