Remove role dispatch optimization
The MoarVM dispatchers included a hook to bypass the `find_method` call
on the metaclass object of a role. That optimization was possible because
the behavior of the thunk returned by that `find_method` is static:
discard the invocant and just call the method on the pun.

But that behavior is wrong if the invocant is containerized or an empty
container that defaults to the role. In such a case the role's method
should be able to store something in the container. So the thunk should
be adjusted to take the containerization status of the invocant into
account. This means that the behavior will no longer be static and the
dispatchers optimization will no longer be valid. Therefore the
dispatchers optimization was removed and `find_method` was slightly
optimized instead.

This change does have a negative impact on the performance of method calls
on a role. On the whole there doesn't appear to be a big impact on the
performance of Rakudo on MoarVM, possibly because all other method lookup
becomes a bit faster and because the core often uses the pun directly. On
other VMs, the overall performance may well be improved by the slight
optimization of `find_method`.