Fix delegation if method is called on containerized role
If a method is called on a role, the call is delegated to the role's pun.
This delegation works fine in case of a _bare_ role, but it did not take
into account the containerization status of the invocant argument.

If the invocant argument is a container, the role's method may expect to
be able to store something in the container, for example to support
auto-vivification. In this case, the method call will now be delegated to
a Proxy that fetches the pun until something is stored in the container.