Fix cases where DESTROY is invoked on its own stack
Finalization code could either be invoked as part of an existing call
stack or by a dedicated GC thread. In the latter case it didn't have
`$*STACK-ID` set. Now we take care of it.

Because finalization is handled by the bootstrap, it doesn't have access
to `Rakudo::Internals` to use `NEXT-ID` method. By keeping in mind that
it might be not the only case when `$*STACK-ID` would need to be set by
NQP code, method `next_id` has been added to
`Perl6::Metamodel::Configuration`. It uses `Rakudo::Internals` when
available and falls back to producing negative IDs otherwise. The method
is thread-safe on fallback path by employing NQPLock.