Get back to using Lock in Stash and CompUnit
Use of `Lock::Soft` caused some expected, though not critical,
performance regression. Yet, in the light of a recent discussion on IRC,
the non-blocking locking doesn't outweight the regression. Therefore it
is considered reasonable to get back to use of `Lock`.

`Lock::Soft` as such remains in place though as it might be useful in
applications where `Lock::Async` cannot serve as a `Lock` replacement
due to its specifics.