Introduce curried role binding
This is a generic approach to retrieving a candidate role for a curry.
The curry forwards itself back to the role it curries via the new `bind`
metamethod, which must also exist on the candidate itself, marked by
its own metaobject as a return value. A failure to produce a binding can
be indicated by returning the curry itself, as we do in the case of the
parametric role group.

If we have a generic approach to retrieving a candidate, then we can
take advantage of it by making it a delegate for typechecks. The
parametric role group can now be its nonsignatured candidate in its
entirety in the typechecker's eyes, and the curried role can bridge the
gap between the group and its bound member this way, which is now
relevant in all role typechecks in one way or another.

Allows for the rewrite of `S14-roles/generic-subtyping.t` to pass.