Unify the roles metamethods
Roles' `roles` metamethods lack the `:$local` parameter classes come
with. They have parents, so they should be capable of supporting it,
with the one and only exception being curried roles. Their
`@!parent_typecheck_list` optimization is a burden now, beyond the fact
that `add_parent` ensures inheritability at the least, so they can now
have introspectable parents.

If we put more flags in `roles-ordered`, we should at least try to
preserve current performance however possible. The locality check can be
optimized out of a pure role container, as it needs to qualify as an MRO
member in order to have parents and the complete MRO that implies.