Introduce Metamodel::MonicMachine, a type injection
There's common need throughout the MOP to build similar varieties of
typecheck lists, e.g. the RTL or the MRO. This tends to generate lists
of parents or doees of the `:all`, `:transitive`, or MROish breed. There
should be a pretty sizeable payoff performance-wise if some of that can
be eliminated or otherwise reduced to `:local` variants for smaller
allocations, but more importantly, this should translate nicely for
nqp's metamodel.