Base C3MRO on a monic beckoning in lieu of c3_merge
Yoink! The monic machine can cope with the edge cases all this
boilerplate used to deal with on its own. In order to do so, the `mro`
metamethod should at least be aware that the `:concretizations`
parameter can exist, however, and we do so in the first place to prove
that it can optimize and cope with complex type buffers.

The difference between a beckoning and a C3 merge becomes more apparent
at this point; we don't push a parents list to the end of the buffer
before doing the beckoning (equivalent to a merge otherwise), and doing
so will fail pretty quickly.