Redefine ^concretizations in terms of the monic machine
The goal of this is to make this metamethod simple enough to translate
better to nqp's metamodel. As a side effect, this improves the
metamethod's performance: the number of iterations per second that can
be completed of `^concretizations` on a `buf8` pun as measured with
`Gauge` is 66294 at its peak before, 78816 after, a 19% improvement.