Make coercion more easily inlineable
If the type is already correct, then the method is really small and
can be inlined.  However, this does not appear to have any effect on
runtime, as apparently some other optimization is already preventing
.coerce from being called in the case of already having the correct
type.

Apparently the .coerce method is only called with already correct types
during compilation.  In any case, the small size allows it to be inlined
pretty quickly, so in practice this does not add any method call overhead.