Unify the roles metamethods
There is a `:local` parameter on classes' `^roles` that other kinds of
types lack. It can be argued that `Metamodel::RoleContainer` should be
concerned with this parameter, but if we know the order of the parents
and the order of their roles, we know their MRO, thus it's
`Metamodel::MROMember`'s instead. This should help preserve the
performance of this metamethod on roles and allows for nqp's roles to
ignore it altogether in a reasonable way.