Fix subclasses of .Date(Time) coercion
Calling .Date on a subclass of Date, and .DateTime on a subclass
of DateTime, returned the object itself even if it was a subclass.

This commit makes sure that if you call .Date(Time) on a subclass
of Date(Time), you will actually get the appropriate Date(Time)
object.

This should remove the need of subclasses to provide this logic,
such as e.g. in Games::TauStation::DateTime.