Fix an LTA error message on passing Seq to a List parameter
This is the correct solution for #4864. The previous attempt of fixing
binding by changing parameter type `Positional` equality check to `isa`
check resulted in #4948 and that should've been expected.

The problem with `isa` check is that we only give special meaning to the
`Positional` role and must not do so for consuming classes or their
descendants.