Unify the parents metamethods
The problem with `^parents` as opposed to `^roles` is its complexity,
which makes backporting to nqp rather cumbersome. Related to this is a
possible invocation of `^compute_mro` via `^mro` on a metaobject before
it has been composed through `^parents(:all|:excl)`, which results in an
incomplete MRO can be produced, which means `^parents(:all|:excl)` can
get stuck reporting the wrong result until composition-time. This
affects a check for stubbed methods in parents in `&trait_mod:<is>`.
which wasn't being triggered on a type that appears in core itself:
`X::TooLateForREPR`.

The idea's to make `^parents(:all|:excl)` a little more dynamic and
calculate the order for the list of parents returned before composition,
which should only be demanded of a metaobject that carries a final MRO.
This should give `Metamodel::CurriedRoleHOW` a greater level of support
for `Metamodel::MultipleInheritance`, for instance. Because of the
refactoring we do while we're here, it should become a little easier to
backport these changes to nqp.