Simplify native value / container check in binder
All tests check for $got_native, so extract that first.  And then
use a ternary to set up the type of native that didn't match
expectations.  And then return with error if such a type is set.

This is mainly about having to do fewer checks and reducing the
bytecode size.  Sadly nowhere near inlining levels yet.