Make Iterable.hyper|race take undef as a default
This changes the behaviour of .hyper/.race when given an undefined
value for :batch / :degree.  Before, this was an error.  With this
commit, it will assume the default values.

This makes it much easier to let e.g. CLI's manage defaults for
:degree and :batch, without needing to know what the defaults
actually are.  Just accept --degree and --batch without default
values, and pass them on.