Fix discrepancies between unit and block scoped packages for good
Up to now unit scoped packages had one lexical scope less as that scope would
be created by the block but not a pure statementlist. This caused us to lose
protos generated within that unit scope as packages don't generate QAST for
any generated declarations they may get. Since we synthesize a Block for the
packages' body anyway, we may as well do it right and use this block as a
lexical scope.