Ensure that all stubbed code objects handle cloning correctly
When a meta method gets added to the meta class, a role is created and a clone
of the method added to the role. At that point we have not yet called
IMPL-LINK-META-OBJECT on the RakuAST::Method, so the clone handler was not yet
in place and the clone's code object was not marked as stub. So add a clone
handler right away when we stub and overwrite it with a more comprehensive one
when we link.