Implement Haskell's "span" in 6.e
Inspired by https://stackoverflow.com/questions/72809469/in-raku-how-does-one-write-the-equivalent-of-haskells-span-function

This implements both the method as well as the sub version of "span".

For more flexibility, the condition can also be a List of conditions
(in the sub case), or just have multiple conditions (in the method case).

    .say for span * < 10, 2, 2, 2, 5, 5, 7, 13, 9, 6, 2, 20, 4;
    (2 2 2 5 5 7)
    (13 9 6 2 20 4)

    .say for span (* < 10, * < 20), 2, 2, 2, 5, 5, 7, 13, 9, 6, 2, 20, 4;
    (2 2 2 5 5 7)
    (13 9 6 2)
    (20 4)

    .say for (2, 2, 2, 5, 5, 7, 13, 9, 6, 2, 20, 4).span( * < 5, * < 10 );
    (2 2 2)
    (5 5 7)
    (13 9 6 2 20 4)

Span documentation: https://hackage.haskell.org/package/base-4.16.1.0/docs/Prelude.html#v:span