Fix $_ not containing exception in default blocks withing packages
A default handler did not get an implicit block topic ($_). This made
the QAST compiler find the comp unit's $_ lexical instead of the nearest
one. That $_ however did not contain the exception when we are in a
package as package blocks get their own $_. Fix by giving the default
block an implicit topic, defaulting to the $_ of the outer lexical
scope, i.e. the CATCH block's.