Make sure random generator is not used between nqp::rand calls
The RNG engine is also used internally by VM by, say, the hashing
algorithm. If any of the core or `Test` code is somehow uses the RNG,
directly or not, the results of the test cannot be guaranteed. Therefore
we need to make sure that the two `nqp::rand_I` calls take place as
early after `nqp::srand` as possible, and immediately follow one
another.