Add sub version of .head / .skip / .tail (#4973)
As requested in https://github.com/Raku/problem-solving/issues/328

There is one biggish issue with this, as the sub "skip" clashes with
the "skip" sub as exported by Test.  Fixed this by giving the version
of "skip" from Test its own proto.

Since we added 3 more subs to the core, these tests all needed to be
adapted as well.