Introduce "nano" term for 6.e+
Returns nanoseconds since epoch, to allow HLL access to more accurate
timing.

Specifically did not include that for 6.c.  Unfortunately, it's pretty
hard to add support in the grammar for a nano term in 6.e only, so
the older settings create a runtime error.  Suggestions for fixing this
deficiency are welcome.