Implement Haskell's "span" as "snip" in 6.e
Inspired by https://stackoverflow.com/questions/72809469/in-raku-how-does-one-write-the-equivalent-of-haskells-span-function

Changed the name from "span" to "snip" because of slightly extended functionality from "span" because the condition can also be a List of conditions
(in the sub case), or just have multiple conditions (in the method case).

A "snip" in the provided Iterable will occur whenever a smartmatch with the given value returns False.  It will then go on to the next condition (if any).  If no other conditions are given, will produce the rest of the Iterable.

    .say for snip * < 10, 2, 2, 2, 5, 5, 7, 13, 9, 6, 2, 20, 4;
    (2 2 2 5 5 7)
    (13 9 6 2 20 4)

    .say for snip (* < 10, * < 20), 2, 2, 2, 5, 5, 7, 13, 9, 6, 2, 20, 4;
    (2 2 2 5 5 7)
    (13 9 6 2)
    (20 4)

    .say for (2, 2, 2, 5, 5, 7, 13, 9, 6, 2, 20, 4).snip( * < 5, * < 10 );
    (2 2 2)
    (5 5 7)
    (13 9 6 2 20 4)

    .say for (2, 2, 2, 5, 5, "a", "b", "c").snip: Int;
    (2 2 2 5 5)
    (a b c)

    .say snip Int, 2, 2, 2, 5, 5, "a", "b", "c");
    (2 2 2 5 5)
    (a b c)

This implements both the method (on Any-iterable, as well as on Supply) and the sub version of "snip".

Span documentation: https://hackage.haskell.org/package/base-4.16.1.0/docs/Prelude.html#v:span