Add IO::Path.under-version-control method in 6.e
Returns True if the invocant appears to be under version control.
This is currently defined as it having a sibling ".git" directory,
or any of its parents having a ".git" directory.

Can be handy for CLI scripts before they do massive changes on a
file system: being under version control provides a bigger chance of
reversibility if something went wrong.