Get around a problem where a module duplicates method objects
If the method is delegating then metamodel invokes `apply_handles` for
each its copy in the method table causing 'method is already defined'
error for delegations. For now it is only `Method::Also` is known which
does this kind of duplication.

The fix is implemented with an additional `:$handles` argument. If true
then no attempt to call `apply_handles` is done.

Until `Method::Also` incorporates support for this argument a workaround
step is done where we make sure that the method object hasn't been
previously installed in the table. This step is to be eliminated as soon
as use of `:!handles` argument is ensured where necessary.

Also added support for using `handles` on multi-candidates. For the
moment this support is also suceptible to the method object duplication
problem. But there is no simple fix for this, so just hoping for
`:!handles` again.